# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Toke Schools CBT Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login for student or staff (admin auto-detected)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password (works for all roles)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /auth/profile-picture:
    patch:
      operationId: updateProfilePicture
      tags: [auth]
      summary: Update profile picture (base64 data URL)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfilePictureBody"
      responses:
        "200":
          description: Profile picture updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /student/exams/{examId}:
    get:
      operationId: getStudentExam
      tags: [students]
      summary: Get exam details for student (with questions)
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exam with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamWithQuestions"
  /student/exams/{examId}/submit:
    post:
      operationId: submitExam
      tags: [students]
      summary: Submit exam answers
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitExamBody"
      responses:
        "200":
          description: Exam submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamResult"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
    StaffPermissions:
      type: object
      properties:
        manage_exams:
          type: boolean
        view_all_exams:
          type: boolean
        view_all_results:
          type: boolean
        manage_students:
          type: boolean
        reset_student_exam:
          type: boolean
      required: [manage_exams, view_all_exams, view_all_results, manage_students, reset_student_exam]
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [student, staff]
      required: [username, password, role]
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        role:
          type: string
          enum: [student, staff, admin]
        name:
          type: string
          nullable: true
        isDefaultPassword:
          type: boolean
        id:
          type: string
        staffRole:
          type: string
          nullable: true
        permissions:
          $ref: "#/components/schemas/StaffPermissions"
          nullable: true
      required: [token, role, name, isDefaultPassword, id]
    ChangePasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required: [currentPassword, newPassword]
    UpdateProfilePictureBody:
      type: object
      properties:
        profilePicture:
          type: string
          description: Base64 encoded data URL (e.g. data:image/jpeg;base64,...)
      required: [profilePicture]
...
    Question:
      type: object
      properties:
        id:
          type: integer
        examId:
          type: integer
        questionText:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctOption:
          type: string
          enum: [A, B, C, D]
      required: [id, examId, questionText, optionA, optionB, optionC, optionD, correctOption]
    ExamWithQuestions:
      type: object
      properties:
        id:
          type: integer
        subject:
          type: string
        class:
          type: string
        durationMinutes:
          type: integer
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        createdBy:
          type: string
        resultsEnabled:
          type: boolean
          nullable: true
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
      required: [id, subject, class, durationMinutes, createdBy, questions]
...
    SubmitExamBody:
      type: object
      properties:
        answers:
          type: object
          additionalProperties:
            type: string
      required: [answers]
    ExamResult:
      type: object
      properties:
        id:
          type: integer
        score:
          type: integer
        total:
          type: integer
        percentage:
          type: number
        grade:
          type: string
        submittedAt:
          type: string
      required: [id, score, total, percentage, grade, submittedAt]
...