# lib/api-zod/src/generated/types/examWithQuestions.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Toke Schools CBT Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Question } from "./question";

export interface ExamWithQuestions {
  id: number;
  subject: string;
  class: string;
  durationMinutes: number;
  startTime?: string | null;
  endTime?: string | null;
  createdBy: string;
  resultsEnabled?: boolean | null;
  questions: Question[];
}