# lib/api-zod/src/generated/types/updateProfilePictureBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Toke Schools CBT Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProfilePictureBody {
  /** Base64 encoded data URL (e.g. data:image/jpeg;base64,...) */
  profilePicture: string;
}